backend/requirements.txt
fastapi
uvicorn
python-multipart
gspread
google-auth-oauthlib